Address ID converted to Long Added builder for address

+++ inventory-database-service/src/main/java/com/cg/inventorydatabaseservice/entity/Address.java
@@ -28,7 +28,7 @@ import lombok.NoArgsConstructor;
 public class Address {
   @Id
   @JsonIgnore
-  private Integer addressId;
+  private Long addressId;
   @Column(length = 25)
   private String city;
   @Column(length = 25)
